refactor(dataSharing): wire new modules and clean global schema
