+++ opencti-platform/opencti-graphql/src/graphql/schema.js
@@ -60,11 +60,9 @@ import fileResolvers from '../resolvers/file';
 import indexedFileResolvers from '../resolvers/indexedFile';
 import globalTypeDefs from '../../config/schema/opencti.graphql';
 import organizationOrIndividualResolvers from '../resolvers/organizationOrIndividual';
-import taxiiResolvers from '../resolvers/taxii';
-import feedResolvers from '../resolvers/feed';
+
 import taskResolvers from '../resolvers/backgroundTask';
 import retentionResolvers from '../resolvers/retentionRule';
-import streamResolvers from '../resolvers/stream';
 import statusResolvers from '../resolvers/status';
 import ruleResolvers from '../resolvers/rule';
 import stixResolvers from '../resolvers/stix';
@@ -180,9 +178,6 @@ const schemaResolvers = [
   constraintDirectiveTypeDefs,
   // INTERNAL
   globalResolvers,
-  taxiiResolvers,
-  feedResolvers,
-  streamResolvers,
   statusResolvers,
   logResolvers,
   rabbitmqMetricsResolvers,
