+++ opencti-platform/opencti-graphql/src/modules/attributes/internalObject-registrationAttributes.ts
@@ -21,7 +21,6 @@ import {
   ENTITY_TYPE_CAPABILITY,
   ENTITY_TYPE_CONNECTOR,
   ENTITY_TYPE_CONNECTOR_MANAGER,
-  ENTITY_TYPE_FEED,
   ENTITY_TYPE_GROUP,
   ENTITY_TYPE_HISTORY,
   ENTITY_TYPE_MIGRATION_REFERENCE,
@@ -34,9 +33,7 @@ import {
   ENTITY_TYPE_SETTINGS,
   ENTITY_TYPE_STATUS,
   ENTITY_TYPE_STATUS_TEMPLATE,
-  ENTITY_TYPE_STREAM_COLLECTION,
   ENTITY_TYPE_SYNC,
-  ENTITY_TYPE_TAXII_COLLECTION,
   ENTITY_TYPE_USER,
   ENTITY_TYPE_WORK,
 } from '../../schema/internalObject';
@@ -467,68 +464,6 @@ const internalObjectsAttributes: { [k: string]: Array<AttributeDefinition<any>> 
     { name: 'name', label: 'Name', type: 'string', format: 'short', mandatoryType: 'no', editDefault: false, multiple: false, upsert: false, isFilterable: true },
     { name: 'public_key', label: 'PublicKey', type: 'string', format: 'short', mandatoryType: 'external', editDefault: false, multiple: false, upsert: false, isFilterable: true },
     { name: 'last_sync_execution', label: 'Last execution', type: 'date', editDefault: false, mandatoryType: 'no', multiple: false, upsert: false, isFilterable: true },
-  ],
-  [ENTITY_TYPE_TAXII_COLLECTION]: [
-    { name: 'name', label: 'Name', type: 'string', format: 'short', mandatoryType: 'external', editDefault: true, multiple: false, upsert: false, isFilterable: true },
-    { name: 'description', label: 'Description', type: 'string', format: 'text', mandatoryType: 'no', editDefault: false, multiple: false, upsert: true, isFilterable: true },
-    { name: 'filters', label: 'Filters', type: 'string', format: 'text', mandatoryType: 'no', editDefault: false, multiple: false, upsert: false, isFilterable: false },
-    { name: 'taxii_public', label: 'Public taxii', type: 'boolean', mandatoryType: 'external', editDefault: true, multiple: false, upsert: false, isFilterable: true },
-    { name: 'include_inferences', label: 'Include inferences', type: 'boolean', mandatoryType: 'no', editDefault: true, multiple: false, upsert: false, isFilterable: false },
-    { name: 'score_to_confidence', label: 'Copy OpenCTI scores to confidence level for indicators', type: 'boolean', mandatoryType: 'no', editDefault: true, multiple: false, upsert: false, isFilterable: false },
-    authorizedMembers,
-  ],
-  [ENTITY_TYPE_FEED]: [
-    { name: 'name', label: 'Name', type: 'string', format: 'short', mandatoryType: 'external', editDefault: true, multiple: false, upsert: false, isFilterable: true },
-    { name: 'description', label: 'Description', type: 'string', format: 'text', mandatoryType: 'no', editDefault: false, multiple: false, upsert: false, isFilterable: true },
-    { name: 'filters', label: 'Filters', type: 'string', format: 'text', mandatoryType: 'no', editDefault: false, multiple: false, upsert: false, isFilterable: false },
-    { name: 'separator', label: 'Separator', type: 'string', format: 'short', mandatoryType: 'external', editDefault: true, multiple: false, upsert: false, isFilterable: true },
-    { name: 'rolling_time', label: 'Rolling time', type: 'numeric', precision: 'long', mandatoryType: 'external', editDefault: true, multiple: false, upsert: false, isFilterable: true },
-    { name: 'include_header', label: 'Include header', type: 'boolean', mandatoryType: 'external', editDefault: true, multiple: false, upsert: false, isFilterable: true },
-    { name: 'feed_public', label: 'Public feed', type: 'boolean', mandatoryType: 'external', editDefault: true, multiple: false, upsert: false, isFilterable: true },
-    { name: 'feed_types', label: 'Feed types', type: 'string', format: 'short', mandatoryType: 'external', editDefault: true, multiple: true, upsert: false, isFilterable: true },
-    { name: 'feed_date_attribute', label: 'Selected attribute date', type: 'string', format: 'short', mandatoryType: 'external', editDefault: true, multiple: false, upsert: false, isFilterable: true },
-    {
-      name: 'feed_attributes',
-      label: 'Feed attributes',
-      type: 'object',
-      format: 'standard',
-      mandatoryType: 'internal',
-      editDefault: false,
-      multiple: true,
-      upsert: false,
-      isFilterable: true,
-      mappings: [
-        { name: 'attribute', label: 'Attribute', type: 'string', format: 'short', editDefault: false, mandatoryType: 'internal', multiple: false, upsert: true, isFilterable: true },
-        { name: 'multi_match_strategy', label: 'Multi-match strategy', type: 'string', format: 'short', editDefault: false, mandatoryType: 'no', multiple: false, upsert: true, isFilterable: false },
-        { name: 'multi_match_separator', label: 'Multi-match separator', type: 'string', format: 'short', editDefault: false, mandatoryType: 'no', multiple: false, upsert: true, isFilterable: false },
-        {
-          name: 'mappings',
-          label: 'Mappings',
-          type: 'object',
-          format: 'standard',
-          editDefault: false,
-          mandatoryType: 'internal',
-          multiple: true,
-          upsert: true,
-          isFilterable: false,
-          mappings: [
-            { name: 'type', label: 'Type', type: 'string', format: 'short', editDefault: false, mandatoryType: 'internal', multiple: false, upsert: true, isFilterable: false },
-            { name: 'attribute', label: 'Attribute', type: 'string', format: 'short', editDefault: false, mandatoryType: 'internal', multiple: false, upsert: true, isFilterable: false },
-            { name: 'relationship_type', label: 'Relationship type', type: 'string', format: 'short', editDefault: false, mandatoryType: 'no', multiple: false, upsert: true, isFilterable: false },
-            { name: 'target_entity_type', label: 'Target entity type', type: 'string', format: 'short', editDefault: false, mandatoryType: 'no', multiple: false, upsert: true, isFilterable: false },
-          ],
-        },
-      ],
-    },
-    authorizedMembers,
-  ],
-  [ENTITY_TYPE_STREAM_COLLECTION]: [
-    { name: 'name', label: 'Name', type: 'string', format: 'short', mandatoryType: 'external', editDefault: true, multiple: false, upsert: false, isFilterable: true },
-    { name: 'description', label: 'Description', type: 'string', format: 'text', mandatoryType: 'no', editDefault: false, multiple: false, upsert: true, isFilterable: true },
-    { name: 'filters', label: 'Filters', type: 'string', format: 'text', mandatoryType: 'no', editDefault: false, multiple: false, upsert: false, isFilterable: false },
-    { name: 'stream_public', label: 'Public stream', type: 'boolean', mandatoryType: 'no', editDefault: false, multiple: false, upsert: false, isFilterable: true },
-    { name: 'stream_live', label: 'Is live', type: 'boolean', mandatoryType: 'no', editDefault: false, multiple: false, upsert: false, isFilterable: true },
-    authorizedMembers,
   ],
   [ENTITY_TYPE_STATUS_TEMPLATE]: [
     { name: 'name', label: 'Name', type: 'string', format: 'short', mandatoryType: 'external', editDefault: true, multiple: false, upsert: false, isFilterable: true },
