+++ opencti-platform/opencti-graphql/src/modules/index.ts
@@ -53,6 +53,9 @@ import './ingestion/ingestion-taxii';
 import './ingestion/ingestion-csv';
 import './ingestion/ingestion-json';
 import './ingestion/ingestion-taxii-collection';
+import './dataSharing/feed';
+import './dataSharing/taxiiCollection';
+import './dataSharing/streamCollection';
 import './indicator/indicator';
 import './decayRule/decayRule';
 import './decayRule/exclusions/decayExclusionRule';
@@ -154,4 +157,7 @@ import './xtm/hub/xtm-hub-graphql';
 import './metrics/metrics-graphql';
 import './authenticationProvider/authenticationProvider-graphql';
 import './customView/customView-graphql';
+import './dataSharing/taxiiCollection-graphql';
+import './dataSharing/feed-graphql';
+import './dataSharing/streamCollection-graphql';
 // endregion
